+++ front/app/components/admin/GraphCards/StatCard/index.tsx
