+++ front/app/components/admin/GraphCards/StatCard/useStatCard/index.ts
