+++ front/app/components/admin/GraphCards/StatCard/useStatCard/parse.ts
