+++ front/app/components/admin/GraphCards/StatCard/useStatCard/typings.ts
