+++ front/app/modules/commercial/analytics/admin/components/ProjectStatusCard/index.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 
 import { Box } from '@citizenlab/cl2-component-library';
 
-import StatCard from '../StatCard';
-import { StatCardProps } from '../StatCard/useStatCard/typings';
+import StatCard from '../../../../../../components/admin/GraphCards/StatCard';
+import { StatCardProps } from '../../../../../../components/admin/GraphCards/StatCard/useStatCard/typings';
 
 import { projectStatusConfig } from './config';
